Internet Watchdog
Frontier has been unable to fix an intermittant problem with our DSL going down randomly. When it goes down, it stays down until the DSL modem is rebooted. Loss of access to the site when no one is there to reboot is very inconvenient. This script running on an inexpensive NodeMCU solves the problem.

Simple lua script to see if we can resolve a few DNS names. If not, then assume the ISP has crashed, so pull a relay for a few seconds to recycle the DSL modem power.